Add markers (and centering) to map in IncidentInfoFragment
For all responders at an incident (currently using dummy
values to simulate values from database), show markers on
the map and center on a responder's marker when that responder
is clicked in the lists on the bottom of the view.
Currently, to simulate an update from the database, a
longclick on the map will update several responder's info.